requirements.txt
# Optimized Streamlit Analytics Dashboard Dependencies
# Core framework
streamlit>=1.28.0,<2.0.0

# Data processing  
pandas>=2.2.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Google Cloud & BigQuery
google-cloud-bigquery>=3.13.0,<4.0.0
google-auth>=2.0.0,<3.0.0
db-dtypes>=1.0.0,<2.0.0

# Visualization
plotly>=5.17.0,<6.0.0
altair>=5.0.0,<6.0.0

# Optional performance enhancements
pyarrow>=10.0.0  # For better BigQuery performance
watchdog>=3.0.0  # For better file watching
